+++ TFG/src/main/java/com/alvaroorgaz/easyfactura/repository/FacturaRectificadaRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface FacturaRectificadaRepository extends JpaRepository<FacturaRectificada, Integer> {
+public interface FacturaRectificadaRepository extends JpaRepository<FacturaRectificada, Long> {
 
     @Query("SELECT fr FROM FacturaRectificada fr WHERE fr.facturaOriginal.idFactura = :idFactura ORDER BY fr.fecha ASC")
-    List<FacturaRectificada> findByFacturaOriginalId(@Param("idFactura") Integer idFactura);
+    List<FacturaRectificada> findByFacturaOriginalId(@Param("idFactura") Long idFactura);
 
     @Query("SELECT fr FROM FacturaRectificada fr WHERE fr.facturaRectificada.idFactura = :idFactura")
-    FacturaRectificada findByFacturaRectificadaId(@Param("idFactura") Integer idFactura);
+    FacturaRectificada findByFacturaRectificadaId(@Param("idFactura") Long idFactura);
 }
